chore: stringRes 추출, price를 정수형으로 변경

+++ app/src/main/java/nextstep/shoppingcart/model/CartItem.kt
@@ -2,6 +2,6 @@ package nextstep.shoppingcart.model
 
 internal data class CartItem(
     val name: String,
-    val price: Float,
+    val price: Int,
     val imageUrl: String?,
 )
